refactor: asyncio usage of cmd into a module
