+++ test_phile/__init__.py
@@ -2,6 +2,7 @@
 """
 .. automodule:: test_phile.test_PySide2
 .. automodule:: test_phile.test_asyncio
+.. automodule:: test_phile.test_cmd
 .. automodule:: test_phile.test_data
 .. automodule:: test_phile.test_datetime
 .. automodule:: test_phile.test_imapclient
